Reorganise the file structure for required files.
The new file structure uses 'bin', 'lib', include', and 'src'. All
source files are placed within the 'src' directory to separate them from
other details. Headers that need to be seen by the whole project can be
placed in 'include'. The 'include' and 'lib' directories are passed to
the compiler.